\name{clickpaste-package}
\alias{clickpaste-package}
\alias{clickpaste}
\docType{package}
\title{
  create a object belong to click_paste class to send event
}
\description{
  objects of click_paste class have useful methods like get_point, single_click and so on.
}
\details{
  objects of click_paste class have useful methods like get_point, single_click and so on.
}
\author{
  Yu, Qiming.

  Maintainer: Yu, Qiming <805968198@qq.com.cn>
}
\references{
  Windows api, mouse event, keyboard event.
}
\keyword{ package }
